Добавил кнопку отмены добавления пользователя

+++ OnLineShop/Model/Customer.cs
@@ -8,7 +8,7 @@ namespace OnLineShop.Model
         public string Surname { get; set; }
         public string Name { get; set; }
         public string Patronymic { get; set; }
-        public uint PhoneNumber { get; set; }
+        public long PhoneNumber { get; set; }
         public string Email { get; set; }
